requirements.txt
# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.34.0
python-multipart==0.0.20
aiofiles==24.1.0
jinja2==3.1.4

# Computer Vision & Face Recognition
opencv-python==4.10.0.84
face-recognition==1.3.0
deepface==0.0.93
numpy==1.26.4
Pillow==11.0.0
scikit-learn==1.6.0

# Database
sqlalchemy==2.0.36
alembic==1.14.0

# Configuration & Environment
python-dotenv==1.0.1
pyyaml==6.0.2
pydantic==2.10.3
pydantic-settings==2.6.1

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.9.0.post0

# Development
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
